refactor: move eval compare to feature folder
