+++ packages/prediction-app/src/features/evaluation-compare/index.ts
@@ -1,0 +1,1 @@
+export * from './EvaluationCompare'
